requirements.txt
beautifulsoup4
pandas
requests
collections-extended
ipython
curl
pymongo
redis-server
numpy

